requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
cycler==0.12.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.55.3
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
Markdown==3.7
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.0.0
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
urllib3==2.3.0
django-cors-headers==4.0.0
groq==0.13.1